Re-org of connection types
