+++ src/filesystems/qsys/FSUtils.ts
@@ -2,7 +2,7 @@ import path from "path";
 import vscode, { l10n } from "vscode";
 import { VscodeTools } from "../../ui/Tools";
 import IBMi from "../../api/IBMi";
-import { ReconnectMode } from "../../api/configuration/config/ConnectionManager";
+import { ReconnectMode } from "../../typings";
 
 /**
  * Called when a member/streamfile is left open when VS Code is closed and re-opened to reconnect (or not) to the previous IBM i, based on the `autoReconnect` global configuration value.
