+++ src/typings.ts
@@ -1,10 +1,8 @@
 import { CustomUI } from "./webviews/CustomUI";
 import Instance from "./Instance";
-import { Tools } from "./api/Tools";
 import { DeployTools } from "./filesystems/local/deployTools";
 import { ComponentRegistry } from './api/components/manager';
-import { ObjectFilters } from './api/configuration/config/ConnectionManager';
-import { DeploymentMethod, FileError, IBMiMember, IBMiObject, WithPath } from "./api/types";
+import { DeploymentMethod, FileError } from "./api/types";
 import { Ignore } from "ignore";
 import { WorkspaceFolder } from "vscode";
 import { VscodeTools } from "./ui/Tools";
@@ -16,18 +14,6 @@ export interface CodeForIBMi {
   evfeventParser: (lines: string[]) => Map<string, FileError[]>,
   tools: typeof VscodeTools,
   componentRegistry: ComponentRegistry
-}
-
-export interface FilteredItem {
-  filter: ObjectFilters
-}
-
-export interface ObjectItem extends FilteredItem, WithPath {
-  object: IBMiObject
-}
-
-export interface MemberItem extends FilteredItem, WithPath {
-  member: IBMiMember
 }
 
 export interface DeploymentParameters {
