+++ src/ui/views/ConnectionBrowser.ts
@@ -1,10 +1,9 @@
 import vscode from 'vscode';
-import { ConnectionData, Server } from '../../typings';
+import { ConnectionConfig, ConnectionData, Server } from '../../typings';
 
 import { instance } from '../../instantiate';
 import { Login } from '../../webviews/login';
 import IBMi from '../../api/IBMi';
-import { ConnectionConfig, ConnectionManager } from '../../api/configuration/config/ConnectionManager';
 import { deleteStoredPassword, getStoredPassword, setStoredPassword } from '../../config/passwords';
 
 type CopyOperationItem = {
