+++ src/ui/views/ProfilesView.ts
@@ -2,10 +2,9 @@
 import vscode, { l10n, window } from 'vscode';
 import { GetNewLibl } from '../../api/components/getNewLibl';
 import { instance } from '../../instantiate';
-import { Profile } from '../../typings';
+import { ConnectionProfile, Profile } from '../../typings';
 import { CommandProfileUi } from '../../webviews/commandProfile';
 import IBMi from '../../api/IBMi';
-import { ConnectionProfile } from '../../api/configuration/config/ConnectionManager';
 
 export class ProfilesView {
   private _onDidChangeTreeData = new vscode.EventEmitter<vscode.TreeItem | undefined | null | void>();
