+++ src/ui/views/searchView.ts
@@ -1,7 +1,6 @@
 import path from 'path';
 import vscode from "vscode";
-import { SearchHit, SearchHitLine, SearchResults, WithPath } from "../../typings";
-import { DefaultOpenMode } from '../../api/configuration/config/ConnectionManager';
+import { DefaultOpenMode, SearchHit, SearchHitLine, SearchResults, WithPath } from "../../typings";
 
 export function initializeSearchView(context: vscode.ExtensionContext) {
   const searchView = new SearchView();
