+++ src/webviews/filters/index.ts
@@ -1,8 +1,8 @@
 import { CustomUI } from "../CustomUI";
 import { Tools } from "../../api/Tools";
 import { instance } from "../../instantiate";
-import { ObjectFilters } from "../../api/configuration/config/ConnectionManager";
 import IBMi from "../../api/IBMi";
+import { ObjectFilters } from "../../typings";
 
 export async function editFilter(filter?: ObjectFilters, copy = false) {
   const connection = instance.getConnection();
