+++ src/webviews/settings/index.ts
@@ -7,10 +7,9 @@ import * as certificates from "../../debug/certificates";
 import { isSEPSupported } from "../../debug/server";
 import { extensionComponentRegistry } from "../../api/components/manager";
 import { instance } from "../../instantiate";
-import { ConnectionData, Server } from '../../typings';
+import { ConnectionConfig, ConnectionData, Server } from '../../typings';
 import { VscodeTools } from "../../ui/Tools";
 import IBMi from "../../api/IBMi";
-import { ConnectionConfig, ConnectionManager } from "../../api/configuration/config/ConnectionManager";
 import { deleteStoredPassword, getStoredPassword, setStoredPassword } from "../../config/passwords";
 
 const EDITING_CONTEXT = `code-for-ibmi:editingConnection`;
